+++ include/parasol/modules/scintilla.h
@@ -16,7 +16,7 @@ class objScintillaSearch;
 
 // Scintilla Lexers.  These codes originate from the Scintilla library.
 
-enum class SCLEX : LONG {
+enum class SCLEX : int {
    NIL = 0,
    ERRORLIST = 10,
    MAKEFILE = 11,
@@ -88,17 +88,17 @@ DEFINE_ENUM_FLAG_OPERATORS(STF)
 
 namespace sci {
 struct SetFont { CSTRING Face; static const AC id = AC(-1); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct ReplaceText { CSTRING Find; CSTRING Replace; STF Flags; LONG Start; LONG End; static const AC id = AC(-2); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct DeleteLine { LONG Line; static const AC id = AC(-3); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct SelectRange { LONG Start; LONG End; static const AC id = AC(-4); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct InsertText { CSTRING String; LONG Pos; static const AC id = AC(-5); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct GetLine { LONG Line; STRING Buffer; LONG Length; static const AC id = AC(-6); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct ReplaceLine { LONG Line; CSTRING String; LONG Length; static const AC id = AC(-7); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct GotoLine { LONG Line; static const AC id = AC(-8); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct ReplaceText { CSTRING Find; CSTRING Replace; STF Flags; int Start; int End; static const AC id = AC(-2); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct DeleteLine { int Line; static const AC id = AC(-3); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct SelectRange { int Start; int End; static const AC id = AC(-4); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct InsertText { CSTRING String; int Pos; static const AC id = AC(-5); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct GetLine { int Line; STRING Buffer; int Length; static const AC id = AC(-6); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct ReplaceLine { int Line; CSTRING String; int Length; static const AC id = AC(-7); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct GotoLine { int Line; static const AC id = AC(-8); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
 struct TrimWhitespace { static const AC id = AC(-9); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct GetPos { LONG Line; LONG Column; LONG Pos; static const AC id = AC(-10); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct GetPos { int Line; int Column; int Pos; static const AC id = AC(-10); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
 struct ReportEvent { static const AC id = AC(-11); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct ScrollToPoint { LONG X; LONG Y; static const AC id = AC(-12); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct ScrollToPoint { int X; int Y; static const AC id = AC(-12); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
 
 } // namespace
 
@@ -115,19 +115,19 @@ class objScintilla : public Object {
    OBJECTID  SurfaceID;          // Refers to the Surface targeted by the Scintilla object.
    SCIF      Flags;              // Optional flags.
    OBJECTID  FocusID;            // Defines the object that is monitored for user focus changes.
-   LONG      Visible;            // If TRUE, indicates the Scintilla object is visible in the target Surface.
-   LONG      LeftMargin;         // The amount of white-space at the left side of the page.
-   LONG      RightMargin;        // Defines the amount of white-space at the right side of the page.
+   int       Visible;            // If TRUE, indicates the Scintilla object is visible in the target Surface.
+   int       LeftMargin;         // The amount of white-space at the left side of the page.
+   int       RightMargin;        // Defines the amount of white-space at the right side of the page.
    struct RGB8 LineHighlight;    // The colour to use when highlighting the line that contains the user's cursor.
    struct RGB8 SelectFore;       // Defines the colour of selected text.  Supports alpha blending.
    struct RGB8 SelectBkgd;       // Defines the background colour of selected text.  Supports alpha blending.
    struct RGB8 BkgdColour;       // Defines the background colour.  Alpha blending is not supported.
    struct RGB8 CursorColour;     // Defines the colour of the text cursor.  Alpha blending is not supported.
    struct RGB8 TextColour;       // Defines the default colour of foreground text.  Supports alpha blending.
-   LONG      CursorRow;          // The current row of the text cursor.
-   LONG      CursorCol;          // The current column of the text cursor.
+   int       CursorRow;          // The current row of the text cursor.
+   int       CursorCol;          // The current column of the text cursor.
    SCLEX     Lexer;              // The lexer for document styling is defined here.
-   LONG      Modified;           // Returns true if the document has been modified and not saved.
+   int       Modified;           // Returns true if the document has been modified and not saved.
 
    // Action stubs
 
@@ -167,39 +167,39 @@ class objScintilla : public Object {
       struct sci::SetFont args = { Face };
       return(Action(AC(-1), this, &args));
    }
-   inline ERR replaceText(CSTRING Find, CSTRING Replace, STF Flags, LONG Start, LONG End) noexcept {
+   inline ERR replaceText(CSTRING Find, CSTRING Replace, STF Flags, int Start, int End) noexcept {
       struct sci::ReplaceText args = { Find, Replace, Flags, Start, End };
       return(Action(AC(-2), this, &args));
    }
-   inline ERR deleteLine(LONG Line) noexcept {
+   inline ERR deleteLine(int Line) noexcept {
       struct sci::DeleteLine args = { Line };
       return(Action(AC(-3), this, &args));
    }
-   inline ERR selectRange(LONG Start, LONG End) noexcept {
+   inline ERR selectRange(int Start, int End) noexcept {
       struct sci::SelectRange args = { Start, End };
       return(Action(AC(-4), this, &args));
    }
-   inline ERR insertText(CSTRING String, LONG Pos) noexcept {
+   inline ERR insertText(CSTRING String, int Pos) noexcept {
       struct sci::InsertText args = { String, Pos };
       return(Action(AC(-5), this, &args));
    }
-   inline ERR getLine(LONG Line, STRING Buffer, LONG Length) noexcept {
+   inline ERR getLine(int Line, STRING Buffer, int Length) noexcept {
       struct sci::GetLine args = { Line, Buffer, Length };
       return(Action(AC(-6), this, &args));
    }
-   inline ERR replaceLine(LONG Line, CSTRING String, LONG Length) noexcept {
+   inline ERR replaceLine(int Line, CSTRING String, int Length) noexcept {
       struct sci::ReplaceLine args = { Line, String, Length };
       return(Action(AC(-7), this, &args));
    }
-   inline ERR gotoLine(LONG Line) noexcept {
+   inline ERR gotoLine(int Line) noexcept {
       struct sci::GotoLine args = { Line };
       return(Action(AC(-8), this, &args));
    }
    inline ERR trimWhitespace() noexcept {
       return(Action(AC(-9), this, NULL));
    }
-   inline ERR getPos(LONG Line, LONG Column, LONG * Pos) noexcept {
-      struct sci::GetPos args = { Line, Column, (LONG)0 };
+   inline ERR getPos(int Line, int Column, int * Pos) noexcept {
+      struct sci::GetPos args = { Line, Column, (int)0 };
       ERR error = Action(AC(-10), this, &args);
       if (Pos) *Pos = args.Pos;
       return(error);
@@ -207,7 +207,7 @@ class objScintilla : public Object {
    inline ERR reportEvent() noexcept {
       return(Action(AC(-11), this, NULL));
    }
-   inline ERR scrollToPoint(LONG X, LONG Y) noexcept {
+   inline ERR scrollToPoint(int X, int Y) noexcept {
       struct sci::ScrollToPoint args = { X, Y };
       return(Action(AC(-12), this, &args));
    }
@@ -243,19 +243,19 @@ class objScintilla : public Object {
       return ERR::Okay;
    }
 
-   inline ERR setVisible(const LONG Value) noexcept {
+   inline ERR setVisible(const int Value) noexcept {
       if (this->initialised()) return ERR::NoFieldAccess;
       this->Visible = Value;
       return ERR::Okay;
    }
 
-   inline ERR setLeftMargin(const LONG Value) noexcept {
+   inline ERR setLeftMargin(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[32];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setRightMargin(const LONG Value) noexcept {
+   inline ERR setRightMargin(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[27];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
@@ -297,12 +297,12 @@ class objScintilla : public Object {
       return field->WriteValue(target, field, 0x01081300, Value, Elements);
    }
 
-   inline ERR setCursorRow(const LONG Value) noexcept {
+   inline ERR setCursorRow(const int Value) noexcept {
       this->CursorRow = Value;
       return ERR::Okay;
    }
 
-   inline ERR setCursorCol(const LONG Value) noexcept {
+   inline ERR setCursorCol(const int Value) noexcept {
       this->CursorCol = Value;
       return ERR::Okay;
    }
@@ -313,19 +313,19 @@ class objScintilla : public Object {
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setModified(const LONG Value) noexcept {
+   inline ERR setModified(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[17];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setAllowTabs(const LONG Value) noexcept {
+   inline ERR setAllowTabs(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[12];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setAutoIndent(const LONG Value) noexcept {
+   inline ERR setAutoIndent(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[18];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
@@ -337,13 +337,13 @@ class objScintilla : public Object {
       return field->WriteValue(target, field, FD_FUNCTION, &Value, 1);
    }
 
-   inline ERR setFoldingMarkers(const LONG Value) noexcept {
+   inline ERR setFoldingMarkers(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[13];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setLineNumbers(const LONG Value) noexcept {
+   inline ERR setLineNumbers(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[14];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
@@ -355,7 +355,7 @@ class objScintilla : public Object {
       return field->WriteValue(target, field, 0x08800300, to_cstring(Value), 1);
    }
 
-   inline ERR setShowWhitespace(const LONG Value) noexcept {
+   inline ERR setShowWhitespace(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[8];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
@@ -373,19 +373,19 @@ class objScintilla : public Object {
       return field->WriteValue(target, field, 0x08800300, to_cstring(Value), 1);
    }
 
-   inline ERR setSymbols(const LONG Value) noexcept {
+   inline ERR setSymbols(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[28];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setTabWidth(const LONG Value) noexcept {
+   inline ERR setTabWidth(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[25];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
    }
 
-   inline ERR setWordwrap(const LONG Value) noexcept {
+   inline ERR setWordwrap(const int Value) noexcept {
       auto target = this;
       auto field = &this->Class->Dictionary[29];
       return field->WriteValue(target, field, FD_LONG, &Value, 1);
@@ -400,9 +400,9 @@ class objScintilla : public Object {
 // ScintillaSearch methods
 
 namespace ss {
-struct Next { LONG Pos; static const AC id = AC(-1); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct Prev { LONG Pos; static const AC id = AC(-2); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct Find { LONG Pos; STF Flags; static const AC id = AC(-3); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct Next { int Pos; static const AC id = AC(-1); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct Prev { int Pos; static const AC id = AC(-2); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
+struct Find { int Pos; STF Flags; static const AC id = AC(-3); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
 
 } // namespace
 
@@ -416,22 +416,22 @@ class objScintillaSearch : public Object {
    objScintilla * Scintilla;    // Targets a Scintilla object for searching.
    CSTRING Text;                // The string sequence to search for.
    STF     Flags;               // Optional flags.
-   LONG    Start;               // Start of the current/most recent selection
-   LONG    End;                 // End of the current/most recent selection
-   inline ERR next(LONG * Pos) noexcept {
-      struct ss::Next args = { (LONG)0 };
+   int     Start;               // Start of the current/most recent selection
+   int     End;                 // End of the current/most recent selection
+   inline ERR next(int * Pos) noexcept {
+      struct ss::Next args = { (int)0 };
       ERR error = Action(AC(-1), this, &args);
       if (Pos) *Pos = args.Pos;
       return(error);
    }
-   inline ERR prev(LONG * Pos) noexcept {
-      struct ss::Prev args = { (LONG)0 };
+   inline ERR prev(int * Pos) noexcept {
+      struct ss::Prev args = { (int)0 };
       ERR error = Action(AC(-2), this, &args);
       if (Pos) *Pos = args.Pos;
       return(error);
    }
-   inline ERR find(LONG * Pos, STF Flags) noexcept {
-      struct ss::Find args = { (LONG)0, Flags };
+   inline ERR find(int * Pos, STF Flags) noexcept {
+      struct ss::Find args = { (int)0, Flags };
       ERR error = Action(AC(-3), this, &args);
       if (Pos) *Pos = args.Pos;
       return(error);
